revert changes in lib

+++ lib/bloomfilter/filter_test.go
@@ -8,12 +8,12 @@ import (
 )
 
 func TestFilter(t *testing.T) {
-	for _, maxItems := range []int32{1e0, 1e1, 1e2, 1e3, 1e4, 1e5} {
+	for _, maxItems := range []int{1e0, 1e1, 1e2, 1e3, 1e4, 1e5} {
 		testFilter(t, maxItems)
 	}
 }
 
-func testFilter(t *testing.T, maxItems int32) {
+func testFilter(t *testing.T, maxItems int) {
 	r := rand.New(rand.NewSource(int64(0)))
 	f := newFilter(maxItems)
 	items := make(map[uint64]struct{}, maxItems)
@@ -36,7 +36,7 @@ func testFilter(t *testing.T, maxItems int32) {
 	}
 
 	// Verify that the added items exist in f.
-	i := int32(0)
+	i := 0
 	for h := range items {
 		if !f.Has(h) {
 			t.Fatalf("cannot find item %d on iteration %d for maxItems %d", h, i, maxItems)
@@ -74,7 +74,7 @@ func testFilter(t *testing.T, maxItems int32) {
 
 func TestFilterConcurrent(t *testing.T) {
 	concurrency := 3
-	maxItems := int32(10000)
+	maxItems := 10000
 	doneCh := make(chan struct{}, concurrency)
 	f := newFilter(maxItems)
 	for i := range concurrency {
